+++ cannons-bukkit/src/main/java/at/pavlov/cannons/dao/DeleteCannonTask.java
@@ -1,7 +1,7 @@
 package at.pavlov.cannons.dao;
 
 import at.pavlov.cannons.Cannons;
-import at.pavlov.cannons.scheduler.RunnableAsync;
+import at.pavlov.internal.async.RunnableAsync;
 
 import java.sql.Statement;
 import java.util.UUID;
