+++ cannons-bukkit/src/main/java/at/pavlov/cannons/dao/LoadWhitelistTask.java
@@ -3,7 +3,7 @@ package at.pavlov.cannons.dao;
 import at.pavlov.cannons.Cannons;
 import at.pavlov.cannons.cannon.Cannon;
 import at.pavlov.cannons.cannon.CannonManager;
-import at.pavlov.cannons.scheduler.RunnableAsync;
+import at.pavlov.internal.async.RunnableAsync;
 
 import java.sql.ResultSet;
 import java.sql.Statement;
