+++ cannons-bukkit/src/main/java/at/pavlov/cannons/scheduler/FakeBlockHandler.java
@@ -3,7 +3,6 @@ package at.pavlov.cannons.scheduler;
 import at.pavlov.bukkit.container.BukkitFakeBlockEntry;
 import at.pavlov.cannons.Cannons;
 import at.pavlov.cannons.config.Config;
-import at.pavlov.cannons.dao.AsyncTaskManager;
 import at.pavlov.cannons.dao.MainTaskManager;
 import at.pavlov.internal.enums.FakeBlockType;
 import lombok.Getter;
