+++ cannons-bukkit/src/main/java/at/pavlov/cannons/scheduler/ProjectileObserver.java
@@ -5,7 +5,6 @@ import at.pavlov.bukkit.container.holders.BukkitSoundHolder;
 import at.pavlov.bukkit.projectile.BukkitProjectile;
 import at.pavlov.cannons.Cannons;
 import at.pavlov.cannons.CreateExplosion;
-import at.pavlov.cannons.dao.AsyncTaskManager;
 import at.pavlov.bukkit.projectile.BukkitFlyingProjectile;
 import at.pavlov.cannons.dao.MainTaskManager;
 import at.pavlov.cannons.projectile.ProjectileManager;
